SettingsEditor.populateGrid: remove try/except around creating property value.
The handling of this exception has been broken since f1e94bc42 (almost
four years ago) because the `value` variable used in the except block
no longer exists.

This exception handling was probably meant to catch exceptions when a
value was invalid for a property.  However, the properties are now
constructed without a value and updated later so this is no longer
needed.  In addition, even if an exception was caught, it didn't
`continue` to the next property so it would fail when appending to the
grid of windows.

(cherry picked from commit bffc2aa87b25304da8c3af03ed7b8920dd15cd7c)